bash script for generating classes from templates
